Update head-to-head matchup with logos instead of team names, adjsted all mobile height of popups

+++ FootballDojo/Models/Stats/PlayerStats/PlayerStatLeader.cs
@@ -5,6 +5,6 @@ namespace FootballDojo.Models
         public string Name { get; set; }
         public string Photo { get;set; }
         public string Description { get; set; }
-        public int? Stat {  get; set; }
+        public double? Stat {  get; set; }
     }
 }
